Interface (#1)
Code for the double buffer implementation in interface. The system can be compiled either for AWS (without fpga dummy) or for local (with fpga dummy) using 'make mode=aws' or 'make mode=local'.

Communication is done using POSIX calls with the file 'dram.hex' in local and the real DRAM on AWS. 

Diagrams explaining the system are added.  

The module runs both locally and on AWS without issue.