Cpu main merge (#11)
Adding the cpu branch containing sha_preprocess algorithm, password generator, and cpu_main.cpp, which is the main file for the program. The program has been tested on the AWS cloud system for the softare emulator and works.

Implementations as verification and optimization of the preprocess are still to be made but will be done in a future branch.

Thanks to Adam Wendelin for reviewing our work.